refactor: 지하철 구간 인수 테스트 리팩토링
- 구간 관련 클래스 분리
- 구간 패키지 생성
