+++ src/main/java/nextstep/subway/section/application/SectionMapper.java
@@ -1,0 +1,40 @@
+package nextstep.subway.section.application;
+
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.LineRepository;
+import nextstep.subway.line.exception.LineNotFoundException;
+import nextstep.subway.section.domain.Section;
+import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.domain.StationRepository;
+import nextstep.subway.station.exception.StationNotFoundException;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SectionMapper {
+
+    private final StationRepository stationRepository;
+    private final LineRepository lineRepository;
+
+    public SectionMapper(StationRepository stationRepository, LineRepository lineRepository) {
+        this.stationRepository = stationRepository;
+        this.lineRepository = lineRepository;
+    }
+
+    public Section toSection(Long id, SectionRequest sectionRequest) {
+        Line line = getLine(id);
+        Station upStation = getStation(sectionRequest.getUpStationId());
+        Station downStation = getStation(sectionRequest.getDownStationId());
+        return new Section(line, upStation, downStation, sectionRequest.getDistance());
+    }
+
+    private Station getStation(Long id) {
+        return stationRepository.findById(id)
+                .orElseThrow(() -> new StationNotFoundException(id));
+    }
+
+    private Line getLine(Long id) {
+        return lineRepository.findById(id)
+                .orElseThrow(() -> new LineNotFoundException(id));
+    }
+}
