+++ src/main/java/nextstep/subway/section/domain/Section.java
@@ -1,6 +1,7 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.section.domain;
 
 import nextstep.subway.common.domain.BaseEntity;
+import nextstep.subway.line.domain.Line;
 import nextstep.subway.station.domain.Station;
 
 import javax.persistence.Entity;
