+++ src/main/java/nextstep/subway/section/ui/SectionController.java
@@ -1,0 +1,30 @@
+package nextstep.subway.section.ui;
+
+import nextstep.subway.section.application.SectionService;
+import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.section.dto.SectionResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+
+@RestController
+@RequestMapping("/lines/{id}/sections")
+public class SectionController {
+
+    private final SectionService sectionService;
+
+    public SectionController(final SectionService sectionService) {
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping
+    public ResponseEntity<SectionResponse> createSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
+        SectionResponse sectionResponse = sectionService.saveSection(id, sectionRequest);
+        return ResponseEntity.created(URI.create("/lines/" + id + "/sections/" + sectionResponse.getId())).body(sectionResponse);
+    }
+}
