+++ src/test/java/nextstep/subway/line/LineSteps.java
@@ -18,7 +18,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LineSteps {
 
     private static final String URI_LINES = "/lines";
-    private static final String URI_SECTIONS = "/sections";
     private static final String HEADER_LOCATION = "Location";
 
     public static ExtractableResponse<Response> requestCreateLineDx(ExtractableResponse<Response> upStationResponse,
@@ -103,22 +102,6 @@ public class LineSteps {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> requestCreateSection(ExtractableResponse<Response> lineResponse,
-                                                                     ExtractableResponse<Response> upStationResponse,
-                                                                     ExtractableResponse<Response> downStationResponse) {
-        Map<String, Object> params = makeSectionParams(upStationResponse, downStationResponse);
-        String uri = lineResponse.header(HEADER_LOCATION);
-
-        return RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post(uri + URI_SECTIONS)
-                .then().log().all()
-                .extract();
-    }
-
     public static void assertCreateLine(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
         assertThat(response.header(HEADER_LOCATION)).isNotBlank();
@@ -155,15 +138,6 @@ public class LineSteps {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 
-    public static void assertCreateSection(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-        assertThat(response.header("Location")).isNotBlank();
-    }
-
-    public static void assertCreateSectionFail(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
     public static Map<String, Object> makeLineParams(String color, String name, Long upStationId, Long downStationId, int distance) {
         Map<String, Object> params = new HashMap<>();
         params.put("color", color);
@@ -171,15 +145,6 @@ public class LineSteps {
         params.put("upStationId", upStationId);
         params.put("downStationId", downStationId);
         params.put("distance", distance);
-        return params;
-    }
-
-    private static Map<String, Object> makeSectionParams(ExtractableResponse<Response> upStationResponse,
-                                                         ExtractableResponse<Response> downStationResponse) {
-        Map<String, Object> params = new HashMap<>();
-        params.put("upStationId", upStationResponse.jsonPath().getLong("id"));
-        params.put("downStationId", downStationResponse.jsonPath().getLong("id"));
-        params.put("distance", 1000);
         return params;
     }
 }
